requirements.txt
imgaug==0.4.0
numpy==1.22.3
matplotlib==3.5.2
scipy==1.8.1
scikit-image==0.19.3
Pillow==9.2.0
torch==1.13.0
torchvision==0.14.0
pandas==1.4.2
PyYAML==6.0
timm==0.6.12
